pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "stepback"
version = "0.0.1"
authors = [
  { name="Fabian Schaipp", email="fabian.schaipp@gmail.com" },
  { name="Aaron Mishkin", email="amishkin@cs.stanford.edu" },
]
description = "Code to reproduce experiments from paper 'Level Set Teleportation: An Optimization Perspective' by Aaron Mishkin, Alberto Bietti, and Robert M. Gower."
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy==1.26.4",
    "torch",
    "torchvision",
    "scikit-learn",
    "scipy",
    "typing_extensions",
    "tqdm",
    "matplotlib",
    "pandas",
    "pytorch-minimize",
    "black",
    "flake8",
    "pylint",
    "mypy",
    "requests",
    "experiment_utils @ git+https://github.com/aaronpmishkin/experiment_utils.git@teleport",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
scaffold = ["py.typed"]

[project.urls]
"Homepage" = "https://github.com/aaronpmishkin/teleport"
#"Bug Tracker" = ""
